chore: Move loot-related code into 'loot' package
