+++ src/main/java/org/terasology/drops/loot/LootSystem.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 MovingBlocks
+ * Copyright 2020 MovingBlocks
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.drops;
+package org.terasology.drops.loot;
 
 import org.terasology.entitySystem.entity.EntityManager;
 import org.terasology.entitySystem.entity.EntityRef;
@@ -35,8 +35,8 @@ import java.util.List;
 import java.util.Map;
 
 @RegisterSystem
-@Share(LootPool.class)
-public class LootPool extends BaseComponentSystem {
+@Share(LootSystem.class)
+public class LootSystem extends BaseComponentSystem {
     @In
     private PrefabManager prefabManager;
 
@@ -46,14 +46,16 @@ public class LootPool extends BaseComponentSystem {
     private Map<Name, List<LootableItem>> lootables = new HashMap<>();
     // sum of all frequencies of given group
     private Map<Name, Long> randomThreshold = new HashMap<>();
+
     private Random random = new FastRandom();
+
     @Override
     public void preBegin() {
         final Collection<Prefab> prefabs = prefabManager.listPrefabs(LootableComponent.class);
         for (Prefab prefab : prefabs) {
             final LootableComponent component = prefab.getComponent(LootableComponent.class);
             for (LootableItem entry : component.lootEntries) {
-                if (!lootables.containsKey(entry.group)){
+                if (!lootables.containsKey(entry.group)) {
                     lootables.put(entry.group, new ArrayList<>());
                     randomThreshold.put(entry.group, 0L);
                 }
@@ -66,7 +68,7 @@ public class LootPool extends BaseComponentSystem {
 
     /**
      * Sets seed for RNG to use.
-     *
+     * <p>
      * If you are going for really deterministic results, please use this method before every group of calls
      * to getRandomLoot(), with param being seed modified by some variable, like block position.
      * Otherwise, if there would be unexpected call to getRandomLoot() in meantime, RNG would be in
@@ -75,23 +77,24 @@ public class LootPool extends BaseComponentSystem {
      * @param seed The value to seed RNG with
      * @return This, for method chaining ({@code lootPool.setSeed(seed).getRandomLoot();})
      */
-    public LootPool setSeed(long seed){
+    public LootSystem setSeed(long seed) {
         this.random = new FastRandom(seed);
         return this;
     }
 
     /**
      * Returns a random {@code LootableItem} from given group.
+     *
      * @param group Group of lootables to retrieve item from
      * @return Random {@code LootableItem}
      */
-    public LootableItem getRandomLoot(@Nonnull String group){
+    public LootableItem getRandomLoot(@Nonnull String group) {
         // When requesting group for which no items have been defined, return something from the default group
         if (!lootables.containsKey(group)) {
             group = "general";
         }
         long randomNumber = Math.abs(random.nextLong()) % randomThreshold.get(group);
-        for (LootableItem item : lootables.get(group)){
+        for (LootableItem item : lootables.get(group)) {
             if (randomNumber < item.frequency) {
                 return item;
             } else {
@@ -104,21 +107,23 @@ public class LootPool extends BaseComponentSystem {
 
     /**
      * Returns a random {@code LootableItem} from the "general" group
+     *
      * @return Random {@code LootableItem}
      */
-    public LootableItem getRandomLoot(){
+    public LootableItem getRandomLoot() {
         return getRandomLoot("general");
     }
 
     /**
      * Converts given {@code LootableItem} to a list of entityRefs of the given LootableItem
+     *
      * @param item LootableItem to convert
      * @return List of entityRefs representing the lootableItem
      */
-    public List<EntityRef> toEntityList(LootableItem item){
+    public List<EntityRef> toEntityList(LootableItem item) {
         List<EntityRef> list = new ArrayList<>(item.maxAmount);
         int amount = random.nextInt(item.minAmount, item.maxAmount);
-        for (int i = 0; i < amount; i++){
+        for (int i = 0; i < amount; i++) {
             list.add(entityManager.create(item.prefab));
         }
         return list;
