+++ src/main/java/org/terasology/drops/loot/LootableComponent.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.drops;
+package org.terasology.drops.loot;
 
 import org.terasology.entitySystem.Component;
