+++ src/main/java/org/terasology/drops/loot/LootableItem.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 MovingBlocks
+ * Copyright 2020 MovingBlocks
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.drops;
+package org.terasology.drops.loot;
 
 import org.terasology.entitySystem.prefab.Prefab;
 import org.terasology.naming.Name;
